// foodstore/src/types/product.ts
import type { ICategory } from "./categoria";

export interface Product {
  id: number;
  eliminado: boolean;
  createdAt: string;
  nombre: string;
  precio: number;
  descripcion: string;
  stock: number;
  imagen: string;
  disponible: boolean;
  categorias: ICategory[];
}

// Para el carrito
export interface CartItem extends Pick<Product, "id" | "nombre" | "precio"> {
  cantidad: number;
}